fabricdemo/graph-query-api/pyproject.toml
[project]
name = "graph-query-api"
version = "0.5.0"
description = "Micro-API for executing graph (GQL) and telemetry (KQL) queries for Foundry agents"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "azure-identity>=1.19.0",
    "azure-cosmos>=4.9.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
    "azure-mgmt-cosmosdb>=9.0.0",
    "sse-starlette>=1.6.0",
    "azure-storage-blob>=12.19.0",
    "azure-search-documents>=11.6.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = []
